requirements.txt
flask==2.3.3
werkzeug==2.3.7
PyGithub==1.55
python-dotenv==0.19.0
gunicorn==21.2.0
